update and add api copy message group, message bag

+++ app/controllers/api/v1/message_managements/message_groups_controller.rb
@@ -31,7 +31,19 @@ class Api::V1::MessageManagements::MessageGroupsController < ApplicationControll
     message_group = MessageGroup.find_by(id: params[:id])
     return render json: {code: 2, message: "Cannot find message group"}  if message_group.blank?
     return render json: {code: 2, message: "User can't permission"} if message_group.user_id != current_user.id
+    return render json: {code: 2, message: "Cannot delete message group"} unless MessageGroupForm.new(message_group, current_user).check_delete
     if message_group.destroy
+      render json: {code: 1, message: "Success!"}
+    else
+      render json: {code: 2, message: "Something went wrong!"}
+    end
+  end
+
+  def copy
+    message_group = MessageGroup.find_by(id: params[:id])
+    return render json: {code: 2, message: "Cannot find message group"}  if message_group.blank?
+    return render json: {code: 2, message: "User can't permission"} if message_group.user_id != current_user.id
+    if CopyObject::MessageManager.new(message_group, "group").call
       render json: {code: 1, message: "Success!"}
     else
       render json: {code: 2, message: "Something went wrong!"}
